+++ src/types/index.ts
@@ -13,6 +13,13 @@ export interface StarkWallet {
   starkKeyPair: ec.KeyPair;
 }
 
+export interface L2Signer {
+  signMessage(message: string): Promise<string>;
+  getAddress(): string;
+}
+
+export { Signer as L1Signer } from '@ethersproject/abstract-signer';
+
 export type EthNetwork = 'dev' | 'ropsten' | 'mainnet';
 
 export interface Config {